Move extracted code into util file.
